License change and addition of inverse.
The license was changed from AGPL to GPL; this shouldn't
make any difference and I don't think it would have harmed
to keep it as AGPL but intuitively I think that the AGPL
is for code that is actually related to something that has
networking purposes or is likely to be used on a server
for services to users over a network. Boolean expressions
are just unrelated enough for me to remove the extra
restriction of the AGPL.

Now of course this means that one now could in theory
create a website that uses this library to let users play
with boolean expressions without being bound by the GPL
to make any changes to the code public, but it is my
sincere hope that you will *still* make those changes
available (under the GPL), after all - I give you this
code for free, too. That is the spirit of the GPL.